refactor: move checker type definitions to top-level types folder
